N°9165 - Moved ExecutionLimits
