+++ application/utils.inc.php
@@ -3216,30 +3216,4 @@ TXT
 				return (int)$sLimit;
 		}
 	}
-
-	/**
-	 * Check if execution should stop due to time limit or memory limit
-	 *
-	 * @param int $iMaxTime             Absolute max time in seconds since epoch
-	 * @param int $iMaxMemoryPercent    Percent of memory usage allowed (0-100)
-	 *
-	 * @return bool true if execution should stop
-	 */
-	public static function ShouldStopExecution(int $iMaxTime = 0, int $iMaxMemoryPercent = 100): bool
-	{
-		if (($iMaxTime != 0) && (time() > $iMaxTime)) {
-			\IssueLog::Debug(__METHOD__.' timeout '.time()." (current) > $iMaxTime (max)");
-			return true;
-		}
-
-		$iMaxMemoryPercent = (int)min(max($iMaxMemoryPercent, 0), 100);
-		$iMemory = memory_get_usage(true);
-		$iMaxMemory = self::GetMemoryLimit() * $iMaxMemoryPercent / 100;
-		if ($iMemory > $iMaxMemory) {
-			\IssueLog::Debug(__METHOD__." Memory limit $iMemory (current) > $iMaxMemory (max)");
-			return true;
-		}
-
-		return false;
-	}
 }
